requirements.txt
# requirements.txt
# Python dependencies for AID agent scripts

openai>=1.0.0
python-dotenv>=1.0.0
